+++ src/test/java/at/porscheinformatik/antimapper/TransformToGroupedHashSets.java
@@ -5,6 +5,7 @@ import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.*;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -20,6 +21,15 @@ public class TransformToGroupedHashSets extends AbstractMapperTest
         Map<Character, Set<String>> dtos = MAPPER.transformToGroupedHashSets(null, GROUPER, BOARDING_PASS);
 
         assertThat(dtos, nullValue());
+    }
+
+    @Test
+    public void testNullToGroupedHashSetsOrEmpty()
+    {
+        Map<Character, Set<String>> dtos =
+            MAPPER.transformToGroupedHashSets(null, GROUPER, BOARDING_PASS, Hint.OR_EMPTY);
+
+        assertThat(dtos, is(Collections.emptyMap()));
     }
 
     @Test