+++ src/test/java/at/porscheinformatik/antimapper/TransformToTreeSet.java
@@ -5,6 +5,7 @@ import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.*;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
@@ -19,6 +20,14 @@ public class TransformToTreeSet extends AbstractMapperTest
         Set<String> dtos = MAPPER.transformToTreeSet(null, BOARDING_PASS);
 
         assertThat(dtos, nullValue());
+    }
+
+    @Test
+    public void testNullToTreeSetOrEmpty()
+    {
+        Set<String> dtos = MAPPER.transformToTreeSet(null, BOARDING_PASS, Hint.OR_EMPTY);
+
+        assertThat(dtos, is(Collections.emptySortedSet()));
     }
 
     @Test